Keep model picker overlays lazy and expose translated menu labels
