+++ packages/react/src/components/model-policy-picker.tsx
@@ -1,31 +1,38 @@
 import type { ClientModel, LatencyMode, ReasoningEffort } from "@opengeni/sdk";
 import {
-  CheckIcon,
   ChevronDownIcon,
   ChevronLeftIcon,
   ChevronRightIcon,
   Globe2Icon,
   KeyRoundIcon,
-  SearchIcon,
   ZapIcon,
 } from "lucide-react";
 import { motion, useReducedMotion } from "motion/react";
-import { Popover } from "radix-ui";
-import { useMemo, useState, type CSSProperties, type ReactNode, type SVGProps } from "react";
+import {
+  Suspense,
+  lazy,
+  useMemo,
+  useId,
+  useState,
+  type CSSProperties,
+  type ReactNode,
+  type SVGProps,
+} from "react";
 import { cn } from "../lib/cn";
 import { usePortalTokenSource, usePortalTokenStyle } from "../lib/use-portal-token-style";
 import {
-  coerceReasoningEffortForModel,
-  effortOptionsForModel,
   findPickerRow,
-  groupPickerRowsByBillingClass,
   labelReasoningEffort,
-  payerSummaryForModel,
   projectClientModelRows,
-  runnableLatencyModesForModel,
   type PickerBillingClass,
   type PickerModelRow,
 } from "../model-policy";
+
+const LazyModelPolicyPickerMenu = lazy(() =>
+  import("./model-policy-picker-menu").then(({ ModelPolicyPickerPopover }) => ({
+    default: ModelPolicyPickerPopover,
+  })),
+);
 
 type ClientPickerModelRow = PickerModelRow<ClientModel>;
 
@@ -37,6 +44,14 @@ export type ModelPolicyPickerMessages = {
   fast: string;
   fastRateHint: string;
   codexOnly: string;
+  searchLabel?: string;
+  searchPlaceholder?: string;
+  currentModel?: string;
+  noMatches?: string;
+  unsupportedAttachments?: string;
+  thinkingEffort?: string;
+  selected?: string;
+
   billingHints: Record<PickerBillingClass, string>;
 };
 
@@ -48,6 +63,15 @@ export const defaultModelPolicyPickerMessages: ModelPolicyPickerMessages = {
   fast: "Fast",
   fastRateHint: "2× rate",
   codexOnly: "Codex-only session",
+  searchLabel: "Search models or providers",
+  searchPlaceholder: "Search models or providers…",
+  currentModel: "Current model",
+  noMatches: "No matching models. Try a model or provider name.",
+  unsupportedAttachments:
+    "Unsupported attachments stay in the session but are hidden from this model.",
+  thinkingEffort: "Thinking effort",
+  selected: "Selected",
+
   billingHints: {
     opengeni_credits: "Will use credits",
     external: "Provider terms and limits apply",
@@ -211,7 +235,7 @@ function applyCodexOnly(
   );
 }
 
-function effectiveRows(props: ModelPolicyPickerProps): ClientPickerModelRow[] {
+export function effectiveRows(props: ModelPolicyPickerProps): ClientPickerModelRow[] {
   const rows = props.rows !== undefined ? props.rows : projectClientModelRows(props.models ?? []);
   const messages = { ...defaultModelPolicyPickerMessages, ...props.messages };
   return applyCodexOnly(rows, props.codexOnly === true, messages.codexOnly);
@@ -302,200 +326,6 @@ export function PickerAnimatedPage(props: {
   );
 }
 
-/** Model selection stays flat; reasoning never becomes a navigation destination. */
-export function ModelPolicyPickerMenu(props: ModelPolicyPickerProps) {
-  const messages = { ...defaultModelPolicyPickerMessages, ...props.messages };
-  const [query, setQuery] = useState("");
-  const rows = effectiveRows(props);
-  const selected = findPickerRow(rows, props.model);
-  const words = query.toLowerCase().trim().split(/\s+/).filter(Boolean);
-  const filtered = rows.filter((row) => {
-    const text =
-      `${row.label} ${row.id} ${row.providerLabel} ${row.billingClassLabel} ${payerSummaryForModel(row.catalog)}`.toLowerCase();
-    return words.every((word) => text.includes(word));
-  });
-  const groups = groupPickerRowsByBillingClass(filtered)
-    .map((group) => ({
-      ...group,
-      rows: group.rows.filter((row) => words.length > 0 || row.id !== selected?.id),
-    }))
-    .filter((group) => group.rows.length > 0);
-  const choose = (row: ClientPickerModelRow) => {
-    if (!row.selectable || props.disabled) return;
-    if (row.id !== props.model) props.onModelChange(row.id);
-    const effort = coerceReasoningEffortForModel(row.catalog, props.effort);
-    // Hosts may commit the combined model/effort draft through this callback.
-    props.onEffortChange(effort);
-    if (
-      props.latencyMode !== "standard" &&
-      !runnableLatencyModesForModel(row.catalog).includes(props.latencyMode)
-    ) {
-      props.onLatencyModeChange("standard");
-    }
-    props.onOpenChange?.(false);
-  };
-  const modelRow = (row: ClientPickerModelRow) => (
-    <PickerNavRow
-      key={row.id}
-      label={row.label}
-      hint={`${payerSummaryForModel(row.catalog)}${row.unavailableReason ? ` · ${row.unavailableReason}` : ""}`}
-      disabled={props.disabled || !row.selectable}
-      title={row.unavailableReason ?? undefined}
-      active={row.id === props.model}
-      showChevron={false}
-      trailing={
-        row.id === props.model ? <CheckIcon className="size-3.5" aria-label="Selected" /> : null
-      }
-      testId={`model-picker-choice-${row.id}`}
-      onClick={() => choose(row)}
-    />
-  );
-  return (
-    <div
-      data-testid="model-picker-menu"
-      className="flex min-h-0 flex-col"
-      onKeyDown={(event) => {
-        if (
-          !(event.target instanceof HTMLButtonElement) ||
-          !["ArrowDown", "ArrowUp", "Home", "End"].includes(event.key)
-        )
-          return;
-        const buttons = Array.from(
-          event.currentTarget.querySelectorAll<HTMLButtonElement>(
-            '[data-testid^="model-picker-choice-"]:not(:disabled)',
-          ),
-        );
-        const index = buttons.indexOf(event.target);
-        if (index < 0) return;
-        event.preventDefault();
-        const next =
-          event.key === "Home"
-            ? 0
-            : event.key === "End"
-              ? buttons.length - 1
-              : (index + (event.key === "ArrowDown" ? 1 : -1) + buttons.length) % buttons.length;
-        buttons[next]?.focus();
-      }}
-    >
-      <div className="flex items-center gap-2 border-b border-og-border px-2 py-2">
-        <SearchIcon className="size-4 shrink-0 text-og-fg-subtle" aria-hidden />
-        <input
-          aria-label="Search models or providers"
-          placeholder="Search models or providers…"
-          value={query}
-          onChange={(event) => setQuery(event.target.value)}
-          onKeyDown={(event) => {
-            if (event.key !== "Escape" && event.key !== "Tab") event.stopPropagation();
-            if (event.key === "ArrowDown") {
-              event.preventDefault();
-              event.currentTarget
-                .closest('[data-testid="model-picker-menu"]')
-                ?.querySelector<HTMLButtonElement>("button:not(:disabled)")
-                ?.focus();
-            }
-          }}
-          className="h-8 min-w-0 flex-1 bg-transparent text-og-menu text-og-fg outline-hidden placeholder:text-og-fg-subtle"
-        />
-      </div>
-      {props.error ? (
-        <p className="px-2 py-2 text-og-control text-og-status-failed" role="alert">
-          {props.error}
-        </p>
-      ) : null}
-      <div
-        className="min-h-0 flex-1 overflow-y-auto overscroll-contain py-1"
-        data-testid="model-picker-models"
-      >
-        {props.loading ? (
-          <p className="px-2 py-3 text-og-control text-og-fg-subtle">{messages.loading}</p>
-        ) : (
-          <>
-            {selected && words.length === 0 ? (
-              <div className="border-b border-og-border/60 pb-1 mb-1">
-                <p className="px-2.5 py-1.5 text-og-control text-og-fg-subtle">Current model</p>
-                {modelRow(selected)}
-              </div>
-            ) : null}
-            {groups.map((group) => (
-              <section key={group.billingClass} aria-label={group.label} className="py-1">
-                <div className="flex items-center gap-2 px-2.5 py-1.5 text-og-control font-medium text-og-fg-subtle">
-                  <BillingClassMark billingClass={group.billingClass} aria-label="" />
-                  {group.label}
-                </div>
-                {group.rows.map(modelRow)}
-              </section>
-            ))}
-            {filtered.length === 0 ? (
-              <p className="px-2 py-4 text-og-control text-og-fg-subtle">
-                {words.length
-                  ? "No matching models. Try a model or provider name."
-                  : messages.noModels}
-              </p>
-            ) : null}
-          </>
-        )}
-      </div>
-      {selected ? (
-        <div className="border-t border-og-border px-2 py-2">
-          {selected.catalog.capabilities?.inputModalities.includes("image") === false ? (
-            <p className="pb-1.5 text-og-control leading-relaxed text-og-fg-subtle">
-              Unsupported attachments stay in the session but are hidden from this model.
-            </p>
-          ) : null}
-          <ModelThinkingControls {...props} />
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function ModelThinkingControls(props: ModelPolicyPickerProps) {
-  const selected = findPickerRow(effectiveRows(props), props.model);
-  if (!selected) return null;
-  const efforts = effortOptionsForModel(selected.catalog);
-  const messages = { ...defaultModelPolicyPickerMessages, ...props.messages };
-  const supportsFast = runnableLatencyModesForModel(selected.catalog).includes("fast");
-  return (
-    <div
-      className="flex items-center gap-2 text-og-control text-og-fg-muted"
-      data-testid="model-picker-reasoning"
-    >
-      <label className="flex min-h-9 items-center gap-2">
-        {messages.thinking}
-        <select
-          aria-label="Thinking effort"
-          value={coerceReasoningEffortForModel(selected.catalog, props.effort)}
-          disabled={props.disabled || !selected.selectable || efforts.length < 2}
-          onChange={(event) => props.onEffortChange(event.target.value as ReasoningEffort)}
-          className="rounded-og-sm bg-og-surface-2 px-2 py-1.5 text-og-fg outline-hidden focus-visible:ring-2 focus-visible:ring-og-accent/40 disabled:opacity-50"
-        >
-          {efforts.map((effort) => (
-            <option key={effort} value={effort}>
-              {labelReasoningEffort(effort)}
-            </option>
-          ))}
-        </select>
-      </label>
-      {supportsFast && props.allowLatencyMode !== false ? (
-        <button
-          type="button"
-          data-testid="model-picker-fast"
-          disabled={props.disabled || !selected.selectable}
-          aria-pressed={props.latencyMode === "fast"}
-          title={`${messages.fast} · ${messages.fastRateHint}`}
-          onClick={() =>
-            props.onLatencyModeChange(props.latencyMode === "fast" ? "standard" : "fast")
-          }
-          className="ml-auto flex min-h-9 items-center gap-1.5 rounded-og-sm px-2 hover:bg-og-surface-2 focus-visible:ring-2 focus-visible:ring-og-accent/40 disabled:opacity-50"
-        >
-          <ZapIcon className={cn("size-3.5", props.latencyMode === "fast" && "fill-current")} />
-          {messages.fast} <span className="text-og-fg-subtle">{messages.fastRateHint}</span>
-        </button>
-      ) : null}
-    </div>
-  );
-}
-
 /** @internal Shared controlled/uncontrolled state seam. */
 export function useModelPolicyPickerState(props: ModelPolicyPickerProps) {
   const [uncontrolledOpen, setUncontrolledOpen] = useState(props.defaultOpen ?? false);
@@ -508,6 +338,7 @@ export function useModelPolicyPickerState(props: ModelPolicyPickerProps) {
 }
 
 export function ModelPolicyPicker(props: ModelPolicyPickerProps) {
+  const contentId = useId();
   const trigger = usePortalTokenSource<HTMLButtonElement>();
   const portalStyle = usePortalTokenStyle(trigger.source);
   const messages = useMemo(
@@ -530,61 +361,59 @@ export function ModelPolicyPicker(props: ModelPolicyPickerProps) {
     );
   }
   return (
-    <Popover.Root open={open} onOpenChange={setOpen}>
-      <Popover.Trigger asChild>
-        <button
-          ref={trigger.ref}
-          type="button"
-          disabled={props.disabled}
-          aria-label={messages.label}
-          className={cn(
-            "og-root og-model-policy-trigger inline-flex h-[var(--og-model-picker-trigger-height)] min-w-0 max-w-64 items-center gap-1 rounded-full border border-transparent px-2.5 text-og-control text-og-fg-muted outline-hidden transition-colors hover:border-og-border hover:bg-og-surface-2 hover:text-og-fg focus-visible:ring-2 focus-visible:ring-og-accent/40 disabled:cursor-not-allowed disabled:opacity-50 max-sm:h-11 max-sm:max-w-[7.5rem] max-sm:px-2",
-            props.className,
-          )}
+    <>
+      <button
+        ref={trigger.ref}
+        type="button"
+        aria-haspopup="dialog"
+        aria-expanded={open}
+        aria-controls={open ? contentId : undefined}
+        data-state={open ? "open" : "closed"}
+        onClick={() => setOpen(!open)}
+        disabled={props.disabled}
+        aria-label={messages.label}
+        className={cn(
+          "og-root og-model-policy-trigger inline-flex h-[var(--og-model-picker-trigger-height)] min-w-0 max-w-64 items-center gap-1 rounded-full border border-transparent px-2.5 text-og-control text-og-fg-muted outline-hidden transition-colors hover:border-og-border hover:bg-og-surface-2 hover:text-og-fg focus-visible:ring-2 focus-visible:ring-og-accent/40 disabled:cursor-not-allowed disabled:opacity-50 max-sm:h-11 max-sm:max-w-[7.5rem] max-sm:px-2",
+          props.className,
+        )}
+      >
+        <BillingClassMark
+          billingClass={selected?.billingClass ?? billingClassForMissingSelection(props.model)}
+          className="text-og-fg"
+        />
+        <span className="og-model-policy-label-full min-w-0 truncate font-medium text-og-fg max-sm:hidden @max-[20rem]/model-controls:hidden">
+          {selected?.label ?? props.model}
+        </span>
+        <span className="og-model-policy-label-short min-w-0 truncate font-medium text-og-fg sm:hidden @max-[20rem]/model-controls:block">
+          {selected?.shortLabel ?? selected?.label ?? props.model}
+        </span>
+        <span
+          className="og-model-policy-effort min-w-0 shrink-[9999] truncate"
+          title={labelReasoningEffort(props.effort)}
         >
-          <BillingClassMark
-            billingClass={selected?.billingClass ?? billingClassForMissingSelection(props.model)}
-            className="text-og-fg"
+          {labelReasoningEffort(props.effort)}
+        </span>
+        {props.latencyMode === "fast" ? (
+          <ZapIcon
+            className="size-3.5 shrink-0 fill-current stroke-current text-og-fg"
+            aria-label={messages.fast}
+            data-testid="model-picker-fast-icon"
           />
-          <span className="og-model-policy-label-full min-w-0 truncate font-medium text-og-fg max-sm:hidden @max-[20rem]/model-controls:hidden">
-            {selected?.label ?? props.model}
-          </span>
-          <span className="og-model-policy-label-short min-w-0 truncate font-medium text-og-fg sm:hidden @max-[20rem]/model-controls:block">
-            {selected?.shortLabel ?? selected?.label ?? props.model}
-          </span>
-          <span
-            className="og-model-policy-effort min-w-0 shrink-[9999] truncate"
-            title={labelReasoningEffort(props.effort)}
-          >
-            {labelReasoningEffort(props.effort)}
-          </span>
-          {props.latencyMode === "fast" ? (
-            <ZapIcon
-              className="size-3.5 shrink-0 fill-current stroke-current text-og-fg"
-              aria-label={messages.fast}
-              data-testid="model-picker-fast-icon"
-            />
-          ) : null}
-          <ChevronDownIcon className="size-3 shrink-0" />
-        </button>
-      </Popover.Trigger>
-      <Popover.Portal>
-        <Popover.Content
-          aria-label={messages.label}
-          align="start"
-          side={props.menuSide ?? "bottom"}
-          sideOffset={8}
-          collisionPadding={12}
-          className={cn(
-            "og-root og-model-policy-menu z-50 flex max-h-[min(32rem,var(--radix-popover-content-available-height))] w-[22rem] max-w-[calc(100vw-16px)] flex-col overflow-hidden rounded-og-lg border border-og-border bg-og-surface-1 p-[var(--og-model-picker-menu-padding)] text-og-fg shadow-og-lg",
-            props.contentClassName,
-          )}
-          style={{ ...portalStyle, ...props.contentStyle }}
-          data-testid="model-picker-content"
-        >
-          <ModelPolicyPickerMenu {...props} onOpenChange={setOpen} />
-        </Popover.Content>
-      </Popover.Portal>
-    </Popover.Root>
+        ) : null}
+        <ChevronDownIcon className="size-3 shrink-0" />
+      </button>
+
+      {open ? (
+        <Suspense fallback={null}>
+          <LazyModelPolicyPickerMenu
+            {...props}
+            anchor={trigger.currentRef}
+            contentId={contentId}
+            portalStyle={portalStyle}
+            onOpenChange={setOpen}
+          />
+        </Suspense>
+      ) : null}
+    </>
   );
 }
