+++ packages/react/src/index.ts
@@ -477,12 +477,12 @@ export type { ModelPickerProps } from "./components/model-picker";
 export {
   BillingClassMark,
   ModelPolicyPicker,
-  ModelPolicyPickerMenu,
   PickerAnimatedPage,
   PickerBackHeader,
   PickerNavRow,
   defaultModelPolicyPickerMessages,
 } from "./components/model-policy-picker";
+export { ModelPolicyPickerMenu } from "./components/model-policy-picker-menu";
 export type {
   ModelPolicyPickerMessages,
   ModelPolicyPickerProps,
